add createPasskeyDeploymentTransaction as a util function
